rewrite DataReader to consume less memory
Change DataReader's API to `each_tag` which yields a tag when it's
available, then feeds one by one to the formatter. This avoids having to
load all tags in memory.

This also changes DataReader to *not load all files from disk* in
recursive mode. Previously, all files were read completely to memory,
even binary ones, to check if they have a ruby shebang. Shebang
detection is now disabled, as executable ruby scripts are usually not
loaded by other scripts as sources of classs/method definitions.

Files are only read recursively if they end with `.rb`, or if
`--all-files` was passed on the command-line.