add `--format` cli option, defaults to Vim format
To match the behavior of ctags, the default format is now Vim, which
goes along well with the default output file `./tags`. If the output
file is changed to `TAGS`, Emacs format is automatically triggered.

This makes `ripper-tags -R` a compatible replacement for ctags in a Ruby
project.

User can use the `--format` option to explicitly choose between:
  - vim
  - emacs
  - json
  - custom (uses our DefaultFormatter)