Drop Ruby 1.8 compatibility
The CI build doesn't pass anymore for Ruby 1.8 because the "ripper" gem
doesn't build on Ubuntu 14.04, whereas it passed on 12.04. I suspect
this might be a problem with an incompatible Bison on the system.

However, since macOS Sierra can't install 1.8 anymore, and we have
problems having CI run against 1.8 as well, let's drop 1.8 compatibility
altogether, since it's becoming hard to keep up the test suite running.

https://travis-ci.org/tmm1/ripper-tags/jobs/271286645